requirements.txt
Django
djangorestframework
python-decouple
django-cors-headers
drf_extra_fields
filetype